+++ CHANGELOG.md
@@ -10,9 +10,9 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
 
 - **Event and scanner data dependency graphs** to ensure better workspace data synchronization.
 
-### 🩹 Fixed
+### 🔧 Changed
 
-- **Correctly handle commas in blocks.** Using commas like `traits = { trait_cautious, inflexible_strategist }` is valid in Clausewitz scripts, as shown by HOI4 vanilla files. It is now handled properly in the script parser and won't cause parse errors.
+- **Parser improvements:** Correctly handle commas as value separators within blocks and omitted closing brackets at the end of files.
 
 ## [v0.22.2] - 2026-06-28 (Hotfix)